Cast NSUIntegers to unsigned long for use with stringWithFormat:

+++ PodioKit/API/PKAppAPI.m
@@ -11,11 +11,11 @@
 @implementation PKAppAPI
 
 + (PKRequest *)requestForAppWithId:(NSUInteger)appId {
-  return [PKRequest requestWithURI:[NSString stringWithFormat:@"/app/%d", appId] method:PKRequestMethodGET];
+  return [PKRequest requestWithURI:[NSString stringWithFormat:@"/app/%ld", (unsigned long)appId] method:PKRequestMethodGET];
 }
 
 + (PKRequest *)requestToInstallAppWithId:(NSUInteger)appId spaceId:(NSUInteger)spaceId {
-  PKRequest *request = [PKRequest requestWithURI:[NSString stringWithFormat:@"/app/%d/install", appId] method:PKRequestMethodPOST];
+  PKRequest *request = [PKRequest requestWithURI:[NSString stringWithFormat:@"/app/%ld/install", (unsigned long)appId] method:PKRequestMethodPOST];
   
   request.body = @{@"space_id": @(spaceId)};
   
@@ -23,7 +23,7 @@
 }
 
 + (PKRequest *)requestForAppsInSpaceWithId:(NSUInteger)spaceId {
-  NSString *uri = [NSString stringWithFormat:@"/app/space/%d/", spaceId];
+  NSString *uri = [NSString stringWithFormat:@"/app/space/%ld/", (unsigned long)spaceId];
   PKRequest *request = [PKRequest requestWithURI:uri method:PKRequestMethodGET];
   
   return request;
@@ -33,7 +33,7 @@
   PKRequest *request = [PKRequest requestWithURI:@"/app/top/" method:PKRequestMethodGET];
   
   if (limit > 0) {
-    [request.parameters setObject:[NSString stringWithFormat:@"%d", limit] forKey:@"limit"];
+    [request.parameters setObject:[NSString stringWithFormat:@"%ld", (unsigned long)limit] forKey:@"limit"];
   }
 
   return request;
